+++ src/synchronization/SendLocalChangesManager.cpp
@@ -1779,7 +1779,7 @@ void SendLocalChangesManager::sendTags()
 
         errorDescription.clear();
         qint32 rateLimitSeconds = 0;
-        qint32 errorCode = qevercloud::EDAMErrorCode::UNKNOWN;
+        qint32 errorCode = static_cast<qint32>(qevercloud::EDAMErrorCode::UNKNOWN);
 
         QString linkedNotebookAuthToken;
         QString linkedNotebookShardId;
@@ -1896,7 +1896,8 @@ void SendLocalChangesManager::sendTags()
                                               linkedNotebookAuthToken);
         }
 
-        if (errorCode == qevercloud::EDAMErrorCode::RATE_LIMIT_REACHED)
+        if (errorCode ==
+            static_cast<qint32>(qevercloud::EDAMErrorCode::RATE_LIMIT_REACHED))
         {
             if (rateLimitSeconds < 0)
             {
@@ -1932,7 +1933,8 @@ void SendLocalChangesManager::sendTags()
             Q_EMIT rateLimitExceeded(rateLimitSeconds);
             return;
         }
-        else if (errorCode == qevercloud::EDAMErrorCode::AUTH_EXPIRED)
+        else if (errorCode ==
+                 static_cast<qint32>(qevercloud::EDAMErrorCode::AUTH_EXPIRED))
         {
             if (!tag.hasLinkedNotebookGuid())
             {
@@ -1966,7 +1968,8 @@ void SendLocalChangesManager::sendTags()
 
             return;
         }
-        else if (errorCode == qevercloud::EDAMErrorCode::DATA_CONFLICT)
+        else if (errorCode ==
+                 static_cast<qint32>(qevercloud::EDAMErrorCode::DATA_CONFLICT))
         {
             QNINFO("Encountered DATA_CONFLICT exception while trying to send "
                    "new and/or modified tags, it means the incremental sync "
@@ -2147,7 +2150,7 @@ void SendLocalChangesManager::sendSavedSearches()
 
         errorDescription.clear();
         qint32 rateLimitSeconds = 0;
-        qint32 errorCode = qevercloud::EDAMErrorCode::UNKNOWN;
+        qint32 errorCode = static_cast<qint32>(qevercloud::EDAMErrorCode::UNKNOWN);
 
         bool creatingSearch = !search.hasUpdateSequenceNumber();
         if (creatingSearch) {
@@ -2161,7 +2164,8 @@ void SendLocalChangesManager::sendSavedSearches()
                                                     rateLimitSeconds);
         }
 
-        if (errorCode == qevercloud::EDAMErrorCode::RATE_LIMIT_REACHED)
+        if (errorCode ==
+            static_cast<qint32>(qevercloud::EDAMErrorCode::RATE_LIMIT_REACHED))
         {
             if (rateLimitSeconds < 0)
             {
@@ -2197,12 +2201,14 @@ void SendLocalChangesManager::sendSavedSearches()
             Q_EMIT rateLimitExceeded(rateLimitSeconds);
             return;
         }
-        else if (errorCode == qevercloud::EDAMErrorCode::AUTH_EXPIRED)
+        else if (errorCode ==
+                 static_cast<qint32>(qevercloud::EDAMErrorCode::AUTH_EXPIRED))
         {
             handleAuthExpiration();
             return;
         }
-        else if (errorCode == qevercloud::EDAMErrorCode::DATA_CONFLICT)
+        else if (errorCode ==
+                 static_cast<qint32>(qevercloud::EDAMErrorCode::DATA_CONFLICT))
         {
             QNINFO("Encountered DATA_CONFLICT exception while "
                    "trying to send new and/or modified saved searches, "
@@ -2304,7 +2310,7 @@ void SendLocalChangesManager::sendNotebooks()
 
         errorDescription.clear();
         qint32 rateLimitSeconds = 0;
-        qint32 errorCode = qevercloud::EDAMErrorCode::UNKNOWN;
+        qint32 errorCode = static_cast<qint32>(qevercloud::EDAMErrorCode::UNKNOWN);
 
         QString linkedNotebookAuthToken;
         QString linkedNotebookShardId;
@@ -2419,7 +2425,8 @@ void SendLocalChangesManager::sendNotebooks()
                                                    linkedNotebookAuthToken);
         }
 
-        if (errorCode == qevercloud::EDAMErrorCode::RATE_LIMIT_REACHED)
+        if (errorCode ==
+            static_cast<qint32>(qevercloud::EDAMErrorCode::RATE_LIMIT_REACHED))
         {
             if (rateLimitSeconds < 0)
             {
@@ -2454,7 +2461,8 @@ void SendLocalChangesManager::sendNotebooks()
             Q_EMIT rateLimitExceeded(rateLimitSeconds);
             return;
         }
-        else if (errorCode == qevercloud::EDAMErrorCode::AUTH_EXPIRED)
+        else if (errorCode ==
+                 static_cast<qint32>(qevercloud::EDAMErrorCode::AUTH_EXPIRED))
         {
             if (!notebook.hasLinkedNotebookGuid()) {
                 handleAuthExpiration();
@@ -2488,7 +2496,8 @@ void SendLocalChangesManager::sendNotebooks()
 
             return;
         }
-        else if (errorCode == qevercloud::EDAMErrorCode::DATA_CONFLICT)
+        else if (errorCode ==
+                 static_cast<qint32>(qevercloud::EDAMErrorCode::DATA_CONFLICT))
         {
             QNINFO("Encountered DATA_CONFLICT exception while "
                    "trying to send new and/or modified notebooks, "
@@ -2674,7 +2683,7 @@ void SendLocalChangesManager::sendNotes()
 
         errorDescription.clear();
         qint32 rateLimitSeconds = 0;
-        qint32 errorCode = qevercloud::EDAMErrorCode::UNKNOWN;
+        qint32 errorCode = static_cast<qint32>(qevercloud::EDAMErrorCode::UNKNOWN);
 
         if (!note.hasNotebookGuid())
         {
@@ -2869,7 +2878,8 @@ void SendLocalChangesManager::sendNotes()
                                                linkedNotebookAuthToken);
         }
 
-        if (errorCode == qevercloud::EDAMErrorCode::RATE_LIMIT_REACHED)
+        if (errorCode ==
+            static_cast<qint32>(qevercloud::EDAMErrorCode::RATE_LIMIT_REACHED))
         {
             if (rateLimitSeconds < 0)
             {
@@ -2904,7 +2914,8 @@ void SendLocalChangesManager::sendNotes()
             Q_EMIT rateLimitExceeded(rateLimitSeconds);
             return;
         }
-        else if (errorCode == qevercloud::EDAMErrorCode::AUTH_EXPIRED)
+        else if (errorCode ==
+                 static_cast<qint32>(qevercloud::EDAMErrorCode::AUTH_EXPIRED))
         {
             if (!notebook.hasLinkedNotebookGuid()) {
                 handleAuthExpiration();
